# public/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FileFlow Project API NEW",
    "version": "1.0.0",
    "description": "API documentation for the FileFlow Project. All endpoints are synchronized with backend routes as of September 27, 2025."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/v1/analytics/plan": {
      "get": {
        "summary": "Get Plan Analytics",
        "description": "Returns plan analytics for the authenticated user.",
        "responses": {
          "200": {
            "description": "Plan analytics returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/analytics/system": {
      "get": {
        "summary": "System Analytics",
        "description": "Returns system analytics data.",
        "responses": {
          "200": {
            "description": "System analytics returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/analytics/performance": {
      "get": {
        "summary": "Performance Analytics",
        "description": "Returns performance analytics data.",
        "responses": {
          "200": {
            "description": "Performance analytics returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/analytics/trends/users": {
      "get": {
        "summary": "User Trends Analytics",
        "description": "Returns user trends analytics data.",
        "responses": {
          "200": {
            "description": "User trends analytics returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
}
}